+++ Underperforming-Students-vs-Teachers/deadlinefighter.cpp
@@ -3,7 +3,7 @@
 const QString DeadlineFighter::DEFAULT_NAME = "Deadline Fighter";
 
 
-DeadlineFighter::DeadlineFighter(QLabel* widget, Row* row, QString name, int maxHp, double skillSpeed, int cost, int damage)
+DeadlineFighter::DeadlineFighter(QLabel* widget, Row* row, QString name, int maxHp, int skillSpeed, int cost, int damage)
     : AttackStudent(widget, row, name, maxHp, skillSpeed, cost, damage)
 {
