+++ Underperforming-Students-vs-Teachers/teacher.cpp
@@ -3,7 +3,7 @@
 #include <algorithm>
 #include <queue>
 
-Teacher::Teacher(QLabel *widget, Row* row, QString name, int maxHp, double speed, int damage) :
+Teacher::Teacher(QLabel *widget, Row* row, QString name, int maxHp, int speed, int damage) :
     Human(widget, row, name, maxHp), speed(speed), damage(damage)
 {
 
@@ -19,11 +19,11 @@ void Teacher::setDamage(int damage) {
     this->damage = damage;
 }
 
-double Teacher::getSpeed() const {
+int Teacher::getSpeed() const {
     return speed;
 }
 
-void Teacher::setSpeed(double speed) {
+void Teacher::setSpeed(int speed) {
     this->speed = speed;
 }
